// src/app/models/jobApplication.ts
export default class JobApplication{
    jobApplicationId: number;
    companyName: string;
    agencyName: string;
    webURL: string;
    contactPersonName: string;
    contactEmail: string;
    phoneNumber: string;
    city: string;
    province: string;
    appliedOn: Date;
    appStatus: string;
    appStatusDisplay: string;
    followUpNotes: string;
}